test: continue coordinator coverage decomposition
